[MoE] Rename mori_prepare_finalize.py -> mori.py, nixl_ep_prepare_finalize.py -> nixl_ep.py
Shorten filenames now that they live in the prepare_finalize/ subdirectory,
following the same convention as naive_dp_ep.py and no_dp_ep.py.
